+++ ush/set_gridparams_GFDLgrid.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import os
-import unittest
 
 from python_utils import (
     import_vars,
@@ -478,28 +477,3 @@ def set_gridparams_GFDLgrid(
         "JEND_OF_RGNL_DOM_WITH_WIDE_HALO_ON_T6SG": jend_of_t7_with_halo_on_t6sg,
     }
 
-
-class Testing(unittest.TestCase):
-    def test_set_gridparams_GFDLgrid(self):
-        grid_params = set_gridparams_GFDLgrid(
-            lon_of_t6_ctr=-97.5,
-            lat_of_t6_ctr=38.5,
-            res_of_t6g=96,
-            stretch_factor=1.4,
-            refine_ratio_t6g_to_t7g=3,
-            istart_of_t7_on_t6g=13,
-            iend_of_t7_on_t6g=84,
-            jstart_of_t7_on_t6g=17,
-            jend_of_t7_on_t6g=80,
-            run_envir="community",
-            verbose=True,
-            nh4=4,
-        )
-
-        self.assertEqual(
-            list(grid_params.values()),
-            [-97.5, 38.5, 216, 192, 6, 1.4, 21, 172, 29, 164],
-        )
-
-    def setUp(self):
-        pass
